Refactor schedule displaying related methods into ScheduleUi class
